requirements.txt
influxdb-client
python-dotenv
pandas
numpy
scipy
matplotlib